+++ src/core/management/commands/fix_contributors.py
@@ -17,7 +17,9 @@ class Command(BaseCommand):
             settings.USER_PREFERENCES_API_BASE,
             settings.USER_PREFERENCES_API_KEY,
         ):
-            raise CommandError('A config parameter is missing in .env! Cannot run command.')
+            raise CommandError(
+                'A config parameter is missing in .env! Cannot run command.'
+            )
 
         for e in progressbar(Entry.objects.all()):
             if e.data and e.data.get('contributors'):
@@ -26,8 +28,10 @@ class Command(BaseCommand):
                 for idx, contributor in enumerate(contributors):
                     if not contributor.get('label'):
                         r = requests.get(
-                            settings.USER_PREFERENCES_API_BASE + f'users/{e.owner.username}/',
+                            settings.USER_PREFERENCES_API_BASE
+                            + f'users/{e.owner.username}/',
                             headers={'X-Api-Key': settings.USER_PREFERENCES_API_KEY},
+                            timeout=settings.REQUESTS_TIMEOUT,
                         )
                         if r.status_code == 200:
                             need_to_save = True
@@ -36,7 +40,9 @@ class Command(BaseCommand):
                             e.data['contributors'][idx]['source'] = e.owner.username
                         else:
                             self.stdout.write(
-                                self.style.WARNING(f'Could not fetch user information for {e.owner.username}')
+                                self.style.WARNING(
+                                    f'Could not fetch user information for {e.owner.username}'
+                                )
                             )
                     roles = contributor.get('roles', [])
                     for role_idx, role in enumerate(roles):
@@ -44,15 +50,28 @@ class Command(BaseCommand):
                             if role.get('source'):
                                 need_to_save = True
                                 _graph, concept = role['source'].rsplit('/', 1)
-                                e.data['contributors'][idx]['roles'][role_idx]['label'] = {
+                                e.data['contributors'][idx]['roles'][role_idx][
+                                    'label'
+                                ] = {
                                     'de': get_preflabel(concept, lang='de'),
                                     'en': titlecase(get_preflabel(concept, lang='en')),
                                 }
                         if role.get('label'):
-                            label_titlecase = titlecase(e.data['contributors'][idx]['roles'][role_idx]['label']['en'])
-                            if label_titlecase != e.data['contributors'][idx]['roles'][role_idx]['label']['en']:
+                            label_titlecase = titlecase(
+                                e.data['contributors'][idx]['roles'][role_idx]['label'][
+                                    'en'
+                                ]
+                            )
+                            if (
+                                label_titlecase
+                                != e.data['contributors'][idx]['roles'][role_idx][
+                                    'label'
+                                ]['en']
+                            ):
                                 need_to_save = True
-                                e.data['contributors'][idx]['roles'][role_idx]['label']['en'] = label_titlecase
+                                e.data['contributors'][idx]['roles'][role_idx]['label'][
+                                    'en'
+                                ] = label_titlecase
                 if need_to_save:
                     e.save()
         self.stdout.write(self.style.SUCCESS('Successfully fixed contributors'))