+++ src/core/schemas/entries/research_project.py
@@ -21,9 +21,13 @@ TYPES = get_collection_members(
 
 class ResearchProjectSchema(BaseSchema):
     project_lead = get_contributors_field_for_role('project_lead', {'order': 1})
-    project_partnership = get_contributors_field_for_role('project_partnership', {'order': 2})
+    project_partnership = get_contributors_field_for_role(
+        'project_partnership', {'order': 2}
+    )
     funding = get_contributors_field_for_role('funding', {'order': 3})
-    funding_category = get_string_field(get_preflabel_lazy('funding_category'), {'order': 4})
+    funding_category = get_string_field(
+        get_preflabel_lazy('funding_category'), {'order': 4}
+    )
     contributors = get_contributors_field({'order': 5})
     date_range = get_date_range_field({'order': 6, 'field_format': 'full'})
     status = get_status_field({'order': 7})