+++ src/media_server/admin.py
@@ -3,8 +3,6 @@ from django.contrib import admin
 from .models import Media
 
 
+@admin.register(Media)
 class MediaAdmin(admin.ModelAdmin):
     pass
-
-
-admin.site.register(Media, MediaAdmin)