+++ src/media_server/management/commands/convert_media.py
@@ -19,13 +19,14 @@ class Command(BaseCommand):
         )
 
     def handle(self, *args, **options):
-
         if not options['media_id']:
             raise CommandError('You have to provide at least one media_id.')
 
         for media_id in options['media_id']:
             if len(media_id) != 22 or not match(r'^[0-9a-zA-Z]{22}$', media_id):
-                raise CommandError(f'This does not look like a valid ShortUUID: {media_id}')
+                raise CommandError(
+                    f'This does not look like a valid ShortUUID: {media_id}'
+                )
 
         for m in progressbar(Media.objects.filter(id__in=options['media_id'])):
             Media.objects.filter(id=m.id).update(status=STATUS_NOT_CONVERTED)